fix(build-logic): don't change build dir in ci
This causes classes compiled via paperweight-userdev to not be added to the classpath correctly, then compilation failed.
